backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database (SQLAlchemy + Async Driver)
sqlalchemy==2.0.25
asyncpg==0.29.0

# Data Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI & Integration
openai==1.10.0
httpx<0.28.0
greenlet

# Telegram
telegram-init-data

# Auth (Phase 3)
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
itsdangerous==2.2.0
email-validator==2.1.0

# Utilities
alembic==1.13.1
